Fixed gh issue #29 error in cast of double to int in FPRangeParameterBinder

+++ repast.simphony.gui/src/repast/simphony/ui/parameters/FPRangeParameterBinder.java
@@ -61,7 +61,7 @@ public class FPRangeParameterBinder extends AbstractParameterBinder {
   @Override
   public void resetToDefault() {
     Object defaultValue = params.getSchema().getDetails(getName()).getDefaultValue();
-    int val = ((Integer) defaultValue).intValue();
+    double val = ((Double) defaultValue).doubleValue();
     spinner.setValue(val);
   }
